Add script to list all occurances of an entity.
Use sqlite3 to list the start and length of all the occurences of
the specified entity from the table occurrence in the
XRAY.entities.{ASIN}.asc file for a book. This aids in extracting
excerpts of these occurrences from the raw output option of
kindleunpack with occ_list.pl or highligthing the excerpt by
modifying the same raw output with occ_bold.pl.