Highlight entity occurrences in kindleunpack rawml.
occ_bold.pl -raw book.rawml < entity.occ > book.html

entity.occ is the output of entity_occ.pl

Options:
-raw rawfile
-out outfile (default STDOUT)
-tag "label string"  (search aid)
-num (precede highligted string with the occ # in [])

The purpose of the occurrence # in [] is to aid knowing
which occurrence is being highligted.

A good label string is the entity name, which will be
enclosed in "[]" to help ensure uniqueness.