Fixes #6 The Martian by Andy Weir
B00EMXBDMA

The occurrences of entity 103, "Arabia" (Saudi Arabia), are really
occurrences of entity 65, "Arabia Terra", a region in the north of
the planet Mars.

The description for entity 16, "Hermes", the spacecraft used in the
novel to shuttle Mars mission crews between Earth orbit and Mars orbit,
is for the canceled European Space Agency spaceplane "Hermes".

Entity 60, the Martian crater Marth, is mis-identified as the lunar crater
Marth.