Fixes #7 The 2020 Commission... by Jeffrey Lewis
The 2020 Commission Report on the North Korean Nuclear Attacks Against
the United States

B079VDR6HM

A number of South Koreans named Park are lumped into xray entity 200
along with the Publisher's address on Park Avenue in NYC on the copyright
page for the book. The entity description is for Park Chung Hee, who is
mentioned once in the book.

Some of the occurrences of entity 94, Major General Lee Jin-won are
actually other characters named Lee.

The character John Sullivan in the book is the acting US secretary of
State, while the xray description of entity 100 is for a John Sullivan
who lived 1740 - 1795.

The xray description for Trump Tower in NYC is for a cancelled project
in Philadelphia.

Entity 68, Ja Song Taek, has only 3 xray references, but is mentioned 42
times in the book.